Solved Migratory Birds from Hackerrank
It is interesting because the first thought is to go with std::map. Then realized that we can't sort the std::map by value and need to go with std::multimap. Even that looked like it couldn't handle duplicate values.
It's interesting because using std::vector was perfectly fine and helped with the requirement of storing just the bird sightings in the vector indices. Also note that constraint is 1-5 which is why we could do this. But if it was like 1 million or something then may be we would need to double check the approach.