Interface, calling of function in multi level inheritance
When an Interface is given and implementation is in library in classes spread across multiple level of inheritance, you create the object using the class that you have implemented in your code deriving from most derived class. But if the implementation is not there in mentioned class, compiler searches the down to top and first available implementation of the function is called.