Find B in sorted matrix
Can you do it in less that quadratic? If yes check all corner cases. Note that the outer while just needs to make sure you are inside the array and when you find the element to get to the lowest pos in that row you need to think about temp.